feat: only export function from user server
