+++ packages/database/src/users/queries.ts
@@ -1,3 +1,5 @@
+"use server"
+
 import { Prisma } from "@prisma/client"
 
 import prisma from "../prisma"
